user_service/requirements.txt
sqlalchemy
grpcio
grpcio-tools
redis
requests